backend/requirements.txt
fastapi
uvicorn[standard]
websockets>=10.4

# ORM y driver de base de datos asíncrono
sqlalchemy[asyncio]
asyncpg
# Usar la versión compilada de psycopg2 para mayor compatibilidad
psycopg2

# Pydantic para validación
pydantic[email]
pydantic-settings

# Celery y Redis
celery>=5.3.0
redis>=4.5.0
aioredis>=2.0.0

# Para variables de entorno
python-dotenv

# Contraseñas
passlib[bcrypt]

# JWT para autenticación
python-jose[cryptography]

# Para procesar datos de formulario (necesario para OAuth2)
python-multipart

# Migraciones de base de datos
alembic 